SeamlessGlasses/requirements.txt
audio_recorder_streamlit
replicate
streamlit
Pillow
opencv-python
pyaudio
numpy
requests
git+https://github.com/speechbrain/speechbrain.git@develop
